% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{combine_replicates}
\alias{combine_replicates}
\alias{dereplicate}
\title{Combine replicates}
\usage{
combine_replicates(metannotate_data_normalized_list)
}
\arguments{
\item{metannotate_data_normalized_list}{List output of \code{\link{normalize}}}
}
\value{
List of two:
\itemize{
\item Tibble of metannotate data, with 'percent_abundance' and 'percent_abundance_sd'
\item Tibble summarizing total normalized counts for genes
}
}
\description{
: Combines replicates in a normalized metannotate table into means with standard deviations
}
