% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\alias{normalize_collapsed_metannotate_data}
\title{Normalize metannotate data}
\usage{
normalize(metannotate_data, normalizing_HMM = "rpoB")
}
\arguments{
\item{metannotate_data}{Tibble of taxonomy-collapsed and mapped metannotate data -
see \code{\link{collapse_by_taxon}}}

\item{normalizing_HMM}{Character vector (length 1) giving the plotting name of an HMM.Family in your table
that you want to normalize all genes within each sample to. This should be a reliable single-copy taxonomic marker gene,
e.g., rpoB, dnaK, and so on.}
}
\value{
List of two:
\itemize{
\item Tibble of metannotate data; 'hits' has now been changed to 'percent_abundance' and is double-normalized
\item Tibble summarizing total normalized counts for genes
}
}
\description{
Double-normlizes metannotate data by HMM length and a single-copy gene marker. The input data has to
be pre-mapped and collapsed by taxonomy
}
