% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{summarize_total_reads_all_genes}
\alias{summarize_total_reads_all_genes}
\title{Counts all HMM hits for all genes in the metannotate dataset}
\usage{
summarize_total_reads_all_genes(metannotate_data, format = "wide", ...)
}
\arguments{
\item{metannotate_data}{Tibble of metannotate data}

\item{format}{Character vector (length 1) of either 'wide' or 'long' (tidyr terminology)
for the style of output table}
}
\value{
Tibble of summarized hit counts, wide format
}
\description{
Counts all HMM hits for all genes in the metannotate dataset
}
\keyword{internal}
